index: Split large raw blobs from main index.
For some repos, having the raw blobs in the shard results in really
large shards 1+GiB. To reduce the loading time of these shards, and
potentially improve caching, we allows splitting the raw data from the
index shard. The current cutoff for an external raw shard is 16MiB, but
likely should be tuned in the future depending on performance. This
makes use of the SHA256 base cache key to deduplicate file blobs between
shards when they are loaded as well as helps check for bad data.

The downside with this change is that the raw data will now include a
vdatble definition for each DocInner where previously this wasn't
duplicated. It might make sense to attach the DocInner to the Doc
directly rather than the raw bytes. This can be done at a later time.