index: Use signed ints across the board.
For better or worse this will reduce the amount of casting needed, and
generally be easier to work with. int64/uint64 are compatible in
flatbuffers, as long at the top bit is not set. It is really hard to
index super large repos at this point, so it is unlikely to imapct
anything.