repo: Allow unreferencing a CREATING shard.
This can happen when two shards are being created at the same time and
the reference resolution process kicks in. The shard that is still in
the process of being written will be in the creating state, and the
resolution process will try to "unreference" it, which is fairly normal
(should be left in the CREATING state).