storage: Start to break storage away from local files.
This was intended to simplify some of the structure around
reading/writing index shards. This initial implementation allows for
loading data from a remote service, it simplified overall deployment on
kubernetes by reducing the number of places a hostPath needed to be
mounted in. In the long run it is entirely possible to see a local disk
access mode come back, along wiht support for things like S3, or GCS, or
other for blob storage.